% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLM}
\alias{BLM}
\title{Contact, threat, and support for Black Lives Matter}
\format{
A data frame with 273 rows and 5 variables:
\describe{
\item{PosCon}{Positive Contact}
\item{NegCon}{Negative Contact}
\item{Real}{Realistic Threat}
\item{Symb}{Symbolic Threat}
\item{blm}{Support for Black Lives Matter}
}
}
\usage{
BLM
}
\description{
A dataset examining predictors of support for BLM
}
\keyword{datasets}
