% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{condom}
\alias{condom}
\title{Inducing Hypocrisy}
\format{
A data frame with 16 rows and 2 variables:
\describe{
\item{CONDITIO}{Condition}
\item{CONDOM}{Condoms purchased}
}
}
\usage{
condom
}
\description{
A dataset examining an intervention to increase condom use intentions
}
\keyword{datasets}
