% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{culture}
\alias{culture}
\title{Measuring Aggression}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{Region}{Region}
\item{Condition}{Condition}
\item{Testosterone}{Testosterone}
}
}
\usage{
culture
}
\description{
A dataset examining a test of the effects of region and insult on aggresion
}
\keyword{datasets}
