% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denial2}
\alias{denial2}
\title{Climate Change Denial}
\format{
A data frame with 1550 rows and 11 variables:
\describe{
\item{CCD}{Climate Change Denial}
\item{ANTIESTABL}{Anti-Establishment Beliefs}
\item{EXCL_ANTIEG}{Anti-Egalitarian Preferences}
\item{TRADVALUE}{Traditional Values}
\item{OPENNESS}{Openness}
\item{PSEUDOSCI}{Pseudoscience beliefs}
\item{AGREEABL}{Agreeableness}
\item{Age}{Age in yeaers}
\item{Gender}{Gender}
\item{Education}{Education = Five Categories}
\item{ed}{Education = Two Categories}
}
}
\usage{
denial2
}
\description{
A dataset examining climate change denial and personality characteristics.
Transgender cases removed
}
\keyword{datasets}
