% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fall18}
\alias{Fall18}
\title{Class Data 2018}
\format{
A data frame with 115 rows and 13 variables:
\describe{
\item{ID}{Identification Number}
\item{Class_Rank}{Factor with four levels. Frosh, Sophmore, Junior, Senior}
\item{GPA}{College Grade Point Average}
\item{Major}{Current Major with two levels. Psychology and Not Psychology}
\item{Psych_Courses}{Number of Psychology Courses Completed}
\item{Study_Hours}{Hours studied per week}
\item{Tv_hours}{Hours watching tv per week}
\item{Summer_Tv}{Hours watching tv per week during Summer}
\item{FirstGen}{Are you a first generation student? Coded as Yes, No}
\item{Siblings}{Number of Siblings}
\item{Grad_School}{Intentions to attend graduate school. Coded as Yes, No}
\item{Expected_Grade}{Grade expected for class}
\item{HS_GPA}{High School Grade Point Average}
}
\usage{
Fall18
}
\description{
Psych 241 Class Data (Fall18)
}
\keyword{datasets}
