% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lab3}
\alias{Lab3}
\title{Lab 3 Data}
\format{
A data frame with 252 rows and 33 variables:
\describe{
\item{weight}{Participant's Weight in lbs}
\item{height}{Participant's Height in inches}
\item{age}{Participants age in years}
}
\usage{
Lab3
}
\description{
Lab 3 Data
}
\keyword{datasets}
