% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yoga}
\alias{yoga}
\title{Yoga for Veterans}
\format{
A data frame with 27 rows and 2 variables:
\describe{
\item{preBDI}{Pre-test depression}
\item{postBDI}{Post-test depression}
}
}
\usage{
yoga
}
\description{
A dataset examining a yoga intervention for vets
}
\keyword{datasets}
